bats/netavark: Run 200-bridge-firewalld tests

+++ tests/containers/bats/netavark.pm
@@ -52,9 +52,6 @@ sub run {
     # Download netavark sources
     patch_sources "netavark", "v$version", "test";
 
-    my $firewalld_backend = script_output "awk -F= '\$1 == \"FirewallBackend\" { print \$2 }' < /etc/firewalld/firewalld.conf";
-    record_info("Firewalld backend", $firewalld_backend);
-
     # Compile helpers & patch tests
     run_command "make examples", timeout => 600;
     if (version->parse(numeric_version($version)) >= version->parse("1.16.0")) {
@@ -63,12 +60,8 @@ sub run {
         run_command "cp target/debug/netavark-connection-tester bin/";
     }
 
-    unless (get_var("RUN_TESTS")) {
-        if ($firewalld_backend ne "iptables") {
-            run_command "rm -f test/100-bridge-iptables.bats";
-            run_command "rm -f test/200-bridge-firewalld.bats";
-        }
-    }
+    # This test was removed on netavark v2.0.0 and the default firewall backend is nftables
+    run_command "rm -f test/100-bridge-iptables.bats" if (version->parse(numeric_version($version)) < version->parse("2.0.0"));
 
     return if check_var("BATS_IGNORE", "all");
     my $errors = run_tests;
